refactor(BE): Github의 정보 받아오는 로직을 Service로 이동 (#58)
