+++ server/src/domain/auth/auth.controller.ts
@@ -1,4 +1,3 @@
-import { HttpService } from '@nestjs/axios';
 import {
   Controller,
   HttpCode,
@@ -10,53 +9,23 @@ import {
   Query,
   UseGuards,
 } from '@nestjs/common';
-import { ConfigService } from '@nestjs/config';
 import { AuthGuard } from '@nestjs/passport';
 import { Request, Response } from 'express';
-import { lastValueFrom } from 'rxjs';
 import { AuthService } from './auth.service';
 
 @Controller()
 export class AuthController {
-  constructor(
-    private readonly authService: AuthService,
-    private readonly configService: ConfigService,
-    private readonly httpService: HttpService,
-  ) {}
+  constructor(private readonly authService: AuthService) {}
 
   @Get('auth/github')
   async authorizeWithGithub(
     @Res({ passthrough: true }) res: Response,
     @Query('code') code: string,
   ) {
-    // TODO Service 계층으로 분리하기
-    const { data } = await lastValueFrom(
-      this.httpService.post('https://github.com/login/oauth/access_token', {
-        client_id: this.configService.get<string>('AUTH_GITHUB_CLIENT_ID'),
-        client_secret: this.configService.get<string>('AUTH_GITHUB_SECRET_KEY'),
-        code: code,
-      }),
-    );
+    const userInfo = await this.authService.getUserInfoUsingGithub(code);
 
-    const githubToken = data.match(/(?<=access_token=).*?(?=&)/)[0];
-    const { data: githubInfo } = await lastValueFrom(
-      this.httpService.get('https://api.github.com/user', {
-        headers: {
-          Authorization: `Bearer ${githubToken}`,
-        },
-      }),
-    );
+    const { email, nickname, avatarUrl } = userInfo;
 
-    const { data: emails } = await lastValueFrom(
-      this.httpService.get('https://api.github.com/user/emails', {
-        headers: {
-          Authorization: `Bearer ${githubToken}`,
-        },
-      }),
-    );
-
-    const { email } = emails.filter((v) => v.primary)[0];
-    const { avatar_url: avatarUrl, login: nickname } = githubInfo;
     let user = await this.authService.findByEmail(email);
     if (user == null) {
       user = await this.authService.join(email, nickname, avatarUrl);
