+++ server/src/domain/auth/auth.service.ts
@@ -3,6 +3,9 @@ import { Repository } from 'typeorm';
 import { InjectRepository } from '@nestjs/typeorm';
 import { JwtService } from '@nestjs/jwt';
 import { User } from '../user/user.entity';
+import { lastValueFrom } from 'rxjs';
+import { ConfigService } from '@nestjs/config';
+import { HttpService } from '@nestjs/axios';
 
 @Injectable()
 export class AuthService {
@@ -10,6 +13,8 @@ export class AuthService {
     @InjectRepository(User)
     private userRepository: Repository<User>,
     private jwtService: JwtService,
+    private readonly configService: ConfigService,
+    private readonly httpService: HttpService,
   ) {}
 
   async join(email: string, nickname: string, avatarUrl: string) {
@@ -61,5 +66,40 @@ export class AuthService {
       nickname: user.nickname,
       profileUrl: user.profileUrl,
     };
+  }
+
+  async getUserInfoUsingGithub(code: string) {
+    const githubToken = await this.getGithubToken(code);
+
+    const { data: githubInfo } = await lastValueFrom(
+      this.httpService.get('https://api.github.com/user', {
+        headers: {
+          Authorization: `Bearer ${githubToken}`,
+        },
+      }),
+    );
+    const { data: emails } = await lastValueFrom(
+      this.httpService.get('https://api.github.com/user/emails', {
+        headers: {
+          Authorization: `Bearer ${githubToken}`,
+        },
+      }),
+    );
+    const { email } = emails.filter((v) => v.primary)[0];
+    const { avatar_url: avatarUrl, login: nickname } = githubInfo;
+    return { email, nickname, avatarUrl };
+  }
+
+  private async getGithubToken(code: string) {
+    const { data } = await lastValueFrom(
+      this.httpService.post('https://github.com/login/oauth/access_token', {
+        client_id: this.configService.get<string>('AUTH_GITHUB_CLIENT_ID'),
+        client_secret: this.configService.get<string>('AUTH_GITHUB_SECRET_KEY'),
+        code: code,
+      }),
+    );
+
+    const githubToken = data.match(/(?<=access_token=).*?(?=&)/)[0];
+    return githubToken;
   }
 }
